+++ src/main/java/com/zerobase/minesweeper/controller/GamerController.java
@@ -77,72 +77,5 @@ public class GamerController {
         return ResponseEntity.ok(HttpStatus.OK);
     }
 
-    /*
-    관리자 페이지 용
-    */
-
-    //회원 목록 보기
-    @PreAuthorize("hasRole('ADMIN')")
-    @GetMapping("/gamers")
-    public ResponseEntity<List<GamersBoardInfoResponse>> getGamers() {
-        return ResponseEntity.ok(gamerService.getGamers().stream()
-                .map(gamerDto -> GamersBoardInfoResponse.builder()
-                        .id(gamerDto.getId())
-                        .name(gamerDto.getName())
-                        .mail(gamerDto.getMail())
-                        .regDt(gamerDto.getRegDt().toLocalDate())
-                        .build())
-                .collect(Collectors.toList()));
-    }
-
-    //특정 회원 보기
-    @PreAuthorize("hasRole('ADMIN')")
-    @GetMapping("/gamer/{gamerId}")
-    public ResponseEntity<GamerInfoResponse> getGamerInfo(@PathVariable("gamerId") @NotBlank Long gamerId) {
-        GamerDto gamerDto = gamerService.getGamerInfo(gamerId);
-        return ResponseEntity.ok(
-                GamerInfoResponse.builder()
-                        .id(gamerDto.getId())
-                        .name(gamerDto.getName())
-                        .mail(gamerDto.getMail())
-                        .isVerified(gamerDto.isVerified())
-                        .regDt(gamerDto.getRegDt())
-                        .verifiedDt(gamerDto.getVerifiedDt())
-                        .build()
-        );
-    }
-
-    //회원 삭제
-    @PreAuthorize("hasRole('ADMIN')")
-    @DeleteMapping("/gamer/{gamerId}")
-    public ResponseEntity<?> deleteGamer(@PathVariable("gamerId") @NotBlank Long gamerId) {
-        gamerService.deleteGamer(gamerId);
-        return ResponseEntity.ok(HttpStatus.OK);
-    }
-
-    //회원 정지
-    @PreAuthorize("hasRole('ADMIN')")
-    @PatchMapping("/gamer/{gamerId}/suspension/{suspend}")
-    public ResponseEntity<?> suspendGamer(@PathVariable("gamerId") @NotBlank Long gamerId,
-                                          @PathVariable("suspend") @NotBlank boolean suspend) {
-        gamerService.suspendGamer(gamerId, suspend);
-        return ResponseEntity.ok(HttpStatus.OK);
-    }
-
-    //회원 검색
-    @PreAuthorize("hasRole('ADMIN')")
-    @GetMapping("/gamer/search")
-    public ResponseEntity<List<GamersBoardInfoResponse>> searchGamers(
-            @RequestParam String keyword
-    ) {
-        return ResponseEntity.ok(gamerService.searchGamers(keyword).stream()
-                .map(gamerDto -> GamersBoardInfoResponse.builder()
-                        .id(gamerDto.getId())
-                        .name(gamerDto.getName())
-                        .regDt(gamerDto.getRegDt().toLocalDate())
-                        .build())
-                .collect(Collectors.toList()));
-    }
-
 }
